+++ vox.Force/fdm_mg_linear_system3.metal
@@ -26,31 +26,31 @@ namespace FdmMgUtils3 {
         
         const array<float, 4> restricted_kernel = {{0.125, 0.375, 0.375, 0.125}};
         
-        array<size_t, 4> kIndices;
+        array<uint, 4> kIndices;
         kIndices[0] = (k > 0) ? 2 * k - 1 : 2 * k;
         kIndices[1] = 2 * k;
         kIndices[2] = 2 * k + 1;
         kIndices[3] = (k + 1 < size_accessor.z) ? 2 * k + 2 : 2 * k + 1;
         
-        array<size_t, 4> jIndices;
+        array<uint, 4> jIndices;
         
-        for (size_t j = jBegin; j < jEnd; ++j) {
+        for (uint j = jBegin; j < jEnd; ++j) {
             jIndices[0] = (j > 0) ? 2 * j - 1 : 2 * j;
             jIndices[1] = 2 * j;
             jIndices[2] = 2 * j + 1;
             jIndices[3] = (j + 1 < size_accessor.y) ? 2 * j + 2 : 2 * j + 1;
             
-            array<size_t, 4> iIndices;
-            for (size_t i = iBegin; i < iEnd; ++i) {
+            array<uint, 4> iIndices;
+            for (uint i = iBegin; i < iEnd; ++i) {
                 iIndices[0] = (i > 0) ? 2 * i - 1 : 2 * i;
                 iIndices[1] = 2 * i;
                 iIndices[2] = 2 * i + 1;
                 iIndices[3] = (i + 1 < size_accessor.x) ? 2 * i + 2 : 2 * i + 1;
                 
                 float sum = 0.0;
-                for (size_t z = 0; z < 4; ++z) {
-                    for (size_t y = 0; y < 4; ++y) {
-                        for (size_t x = 0; x < 4; ++x) {
+                for (uint z = 0; z < 4; ++z) {
+                    for (uint y = 0; y < 4; ++y) {
+                        for (uint x = 0; x < 4; ++x) {
                             float w =
                             restricted_kernel[x] * restricted_kernel[y] * restricted_kernel[z];
                             sum += w * finer(iIndices[x], jIndices[y],
@@ -76,18 +76,18 @@ namespace FdmMgUtils3 {
         ArrayAccessor3<float> coarser(size_accessor, _coarser);
         uint k = id;
         
-        for (size_t j = jBegin; j < jEnd; ++j) {
-            for (size_t i = iBegin; i < iEnd; ++i) {
-                array<size_t, 2> iIndices;
-                array<size_t, 2> jIndices;
-                array<size_t, 2> kIndices;
+        for (uint j = jBegin; j < jEnd; ++j) {
+            for (uint i = iBegin; i < iEnd; ++i) {
+                array<uint, 2> iIndices;
+                array<uint, 2> jIndices;
+                array<uint, 2> kIndices;
                 array<float, 2> iWeights;
                 array<float, 2> jWeights;
                 array<float, 2> kWeights;
                 
-                const size_t ci = i / 2;
-                const size_t cj = j / 2;
-                const size_t ck = k / 2;
+                const uint ci = i / 2;
+                const uint cj = j / 2;
+                const uint ck = k / 2;
                 
                 if (i % 2 == 0) {
                     iIndices[0] = (i > 1) ? ci - 1 : ci;
@@ -125,9 +125,9 @@ namespace FdmMgUtils3 {
                     kWeights[1] = 0.25;
                 }
                 
-                for (size_t z = 0; z < 2; ++z) {
-                    for (size_t y = 0; y < 2; ++y) {
-                        for (size_t x = 0; x < 2; ++x) {
+                for (uint z = 0; z < 2; ++z) {
+                    for (uint y = 0; y < 2; ++y) {
+                        for (uint x = 0; x < 2; ++x) {
                             float w = iWeights[x] * jWeights[y] *
                             kWeights[z] *
                             coarser(iIndices[x], jIndices[y],
